Add rubocop style checks.
Rubocop  statically checks code using the [Ruby Style
Guide](https://github.com/bbatsov/ruby-style-guide). This guide is
opinionated in a lot of regards but it has become the most popular and
agreed upon guide.

I have added a custom rubocop file to tweak some
things to Rails-specific style, and I also ignore checks on some important
files.

All of these can be run with the rubocop gem from the terminal.
Since we have a very large number of corrections, I used rubocop to
create a todo file. This makes it easy to enable style checks in small
chunks so that commits are smaller and easier to follow.